Fixed a bug related to the voltage value data type (unsigned int --> float).

+++ hvsystem.h
@@ -44,7 +44,8 @@ public slots:
     void setPowerChannel(uint8_t nm_chan, bool state);
     void setPowerSystem(bool state);
     void setVoltageChannel(uint8_t nm_chan, ulong voltage);
-    void setVoltageSystem(ulong voltage);
+    void setVoltageSystem(float voltage);
+
